Update so you cannot do commands in commands
If you are in a command, this update prevents you from entering another command while in that message collector. Also, this update adds an protection to see if you actually want to accept that application or not. Also, this update fixed a bug where it checks if an application was already reviewed.